requirements.txt
asgiref
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
pillow
PyJWT
pytz
setuptools
sqlparse
typing_extensions
tzdata
